+++ data/actionsmap/yunohost.yml
@@ -1478,38 +1478,36 @@ tools:
                     extra:
                         pattern: *pattern_port
 
+    subcategories:
 
-#############################
-#         Migrations        #
-#############################
-migrations:
-    category_help: Manage migrations
-    actions:
+      migrations:
+          subcategory_help: Manage migrations
+          actions:
 
-        ### migrations_list()
-        list:
-            action_help: List migrations
-            api: GET /migrations
+              ### tools_migrations_list()
+              list:
+                  action_help: List migrations
+                  api: GET /migrations
 
-        ### migrations_migrate()
-        migrate:
-            action_help: Perform migrations
-            api: POST /migrations/migrate
-            arguments:
-              -t:
-                  help: target migration number (or 0), latest one by default
-                  type: int
-                  full: --target
-              -s:
-                  help: skip the migration(s), use it only if you know what you are doing
-                  full: --skip
-                  action: store_true
+              ### tools_migrations_migrate()
+              migrate:
+                  action_help: Perform migrations
+                  api: POST /migrations/migrate
+                  arguments:
+                    -t:
+                        help: target migration number (or 0), latest one by default
+                        type: int
+                        full: --target
+                    -s:
+                        help: skip the migration(s), use it only if you know what you are doing
+                        full: --skip
+                        action: store_true
 
 
-        ### migrations_state()
-        state:
-            action_help: Show current migrations state
-            api: GET /migrations/state
+              ### tools_migrations_state()
+              state:
+                  action_help: Show current migrations state
+                  api: GET /migrations/state
 
 
 #############################
