+++ src/yunohost/data_migrations/0001_archivemount.py
@@ -1,5 +1,5 @@
 import subprocess
-from yunohost.migrations import Migration
+from yunohost.tools import Migration
 
 
 class MyMigration(Migration):
