+++ src/yunohost/tools.py
@@ -23,6 +23,7 @@
 
     Specific tools
 """
+import re
 import os
 import yaml
 import requests
@@ -33,12 +34,14 @@ import subprocess
 import pwd
 import socket
 from collections import OrderedDict
+from importlib import import_module
 
 import apt
 import apt.progress
 
 from moulinette.core import MoulinetteError, init_authenticator
 from moulinette.utils.log import getActionLogger
+from moulinette.utils.filesystem import read_json, write_to_json
 from yunohost.app import app_fetchlist, app_info, app_upgrade, app_ssowatconf, app_list, _install_appslist_fetch_cron
 from yunohost.domain import domain_add, domain_list, get_public_ip, _get_maindomain, _set_maindomain
 from yunohost.dyndns import dyndns_subscribe
@@ -49,6 +52,7 @@ from yunohost.utils.packages import ynh_packages_version
 
 # FIXME this is a duplicate from apps.py
 APPS_SETTING_PATH= '/etc/yunohost/apps/'
+MIGRATIONS_STATE_PATH = "/etc/yunohost/migrations_state.json"
 
 logger = getActionLogger('yunohost.tools')
 
@@ -622,3 +626,184 @@ def tools_port_available(port):
         return True
     else:
         return False
+
+
+def tools_migrations_list():
+    """
+    List existing migrations
+    """
+
+    migrations = {"migrations": []}
+
+    for migration in _get_migrations_list():
+        migrations["migrations"].append({
+            "number": int(migration.split("_", 1)[0]),
+            "name": migration.split("_", 1)[1],
+            "file_name": migration,
+        })
+
+    return migrations
+
+
+def tools_migrations_migrate(target=None, skip=False):
+    """
+    Perform migrations
+    """
+
+    # state is a datastructure that represents the last run migration
+    # it has this form:
+    # {
+    #     "last_run_migration": {
+    #             "number": "00xx",
+    #             "name": "some name",
+    #         }
+    # }
+    state = tools_migrations_state()
+
+    last_run_migration_number = state["last_run_migration"]["number"] if state["last_run_migration"] else 0
+
+    migrations = []
+
+    # loading all migrations
+    for migration in tools_migrations_list()["migrations"]:
+        logger.debug(m18n.n('migrations_loading_migration',
+            number=migration["number"],
+            name=migration["name"],
+        ))
+
+        try:
+            # this is python builtin method to import a module using a name, we
+            # use that to import the migration as a python object so we'll be
+            # able to run it in the next loop
+            module = import_module("yunohost.data_migrations.{file_name}".format(**migration))
+        except Exception:
+            import traceback
+            traceback.print_exc()
+
+            raise MoulinetteError(errno.EINVAL, m18n.n('migrations_error_failed_to_load_migration',
+                number=migration["number"],
+                name=migration["name"],
+            ))
+            break
+
+        migrations.append({
+            "number": migration["number"],
+            "name": migration["name"],
+            "module": module,
+        })
+
+    migrations = sorted(migrations, key=lambda x: x["number"])
+
+    if not migrations:
+        logger.info(m18n.n('migrations_no_migrations_to_run'))
+        return
+
+    all_migration_numbers = [x["number"] for x in migrations]
+
+    if target is None:
+        target = migrations[-1]["number"]
+
+    # validate input, target must be "0" or a valid number
+    elif target != 0 and target not in all_migration_numbers:
+        raise MoulinetteError(errno.EINVAL, m18n.n('migrations_bad_value_for_target', ", ".join(map(str, all_migration_numbers))))
+
+    logger.debug(m18n.n('migrations_current_target', target))
+
+    # no new migrations to run
+    if target == last_run_migration_number:
+        logger.warn(m18n.n('migrations_no_migrations_to_run'))
+        return
+
+    logger.debug(m18n.n('migrations_show_last_migration', last_run_migration_number))
+
+    # we need to run missing migrations
+    if last_run_migration_number < target:
+        logger.debug(m18n.n('migrations_forward'))
+        # drop all already run migrations
+        migrations = filter(lambda x: target >= x["number"] > last_run_migration_number, migrations)
+        mode = "forward"
+
+    # we need to go backward on already run migrations
+    elif last_run_migration_number > target:
+        logger.debug(m18n.n('migrations_backward'))
+        # drop all not already run migrations
+        migrations = filter(lambda x: target < x["number"] <= last_run_migration_number, migrations)
+        mode = "backward"
+
+    else:  # can't happen, this case is handle before
+        raise Exception()
+
+    # effectively run selected migrations
+    for migration in migrations:
+        if not skip:
+            logger.warn(m18n.n('migrations_show_currently_running_migration', **migration))
+
+            try:
+                if mode == "forward":
+                    migration["module"].MyMigration().migrate()
+                elif mode == "backward":
+                    migration["module"].MyMigration().backward()
+                else:  # can't happen
+                    raise Exception("Illegal state for migration: '%s', should be either 'forward' or 'backward'" % mode)
+            except Exception as e:
+                # migration failed, let's stop here but still update state because
+                # we managed to run the previous ones
+                logger.error(m18n.n('migrations_migration_has_failed', exception=e, **migration), exc_info=1)
+                break
+
+        else:  # if skip
+            logger.warn(m18n.n('migrations_skip_migration', **migration))
+
+        # update the state to include the latest run migration
+        state["last_run_migration"] = {
+            "number": migration["number"],
+            "name": migration["name"],
+        }
+
+    # special case where we want to go back from the start
+    if target == 0:
+        state["last_run_migration"] = None
+
+    write_to_json(MIGRATIONS_STATE_PATH, state)
+
+
+def tools_migrations_state():
+    """
+    Show current migration state
+    """
+    if not os.path.exists(MIGRATIONS_STATE_PATH):
+        return {"last_run_migration": None}
+
+    return read_json(MIGRATIONS_STATE_PATH)
+
+
+def _get_migrations_list():
+    migrations = []
+
+    try:
+        import data_migrations
+    except ImportError:
+        # not data migrations present, return empty list
+        return migrations
+
+    migrations_path = data_migrations.__path__[0]
+
+    if not os.path.exists(migrations_path):
+        logger.warn(m18n.n('migrations_cant_reach_migration_file', migrations_path))
+        return migrations
+
+    for migration in filter(lambda x: re.match("^\d+_[a-zA-Z0-9_]+\.py$", x), os.listdir(migrations_path)):
+        migrations.append(migration[:-len(".py")])
+
+    return sorted(migrations)
+
+
+class Migration(object):
+    def migrate(self):
+        self.forward()
+
+    def forward(self):
+        raise NotImplementedError()
+
+    def backward(self):
+        pass
